refactor(cli)!: remove unused `name` field
It was originally put in to not force generated projects directories to
be named the same as the `artifactId`, but since that is rarely, if
ever, the case, and because the goal of this project is to be
opinionated, it was decided to remove the argument and just use
`artifact_id` for the project directory names instead.

BREAKING CHANGE: The `name` field has been removed from the `Cli`
struct, which could affect external scripts or tools that pass this
argument.