feat!: output rendered files as markdown
It was getting hard to identify which parts of the output indicated
which file it was supposed to be rendered to. This change makes the
program output a markdown document with a section for each rendered
file. This is still just temporary until the program can actually create
the files.

BREAKING CHANGE: This would probably break anything that relies on the
previous output of the program.